Refactor LSP server: modularize helpers, fallbacks, and handlers
Extracted large chunks of logic from `crates/perl-parser/src/lsp/server_impl/mod.rs` into more cohesive modules:
- Moved text processing helpers to `crates/perl-parser/src/lsp/utils.rs`.
- Moved text-based fallback implementations to `crates/perl-parser/src/lsp/fallback/text.rs`.
- Moved non-blocking fallback handlers (`on_definition`, `on_references`, `on_folding_range`) to their respective language handler files (`navigation.rs`, `references.rs`, `symbols.rs`).
- Exposed necessary `LspServer` helper methods as `pub(crate)` to support the new module structure.
- Removed duplicate folding range extraction logic in favor of the unified fallback implementation.
